requirements.txt
starlette
pydantic
fastapi[standard]
sqlmodel
python-dotenv